+++ src/mdxGlobalComponents.js
@@ -97,6 +97,7 @@ import { Structure } from './components/Structure'
 import { TableOfContents } from './components/TableOfContents'
 import { TeamQuote } from './components/TeamQuote'
 import { Tooltip } from './components/Tooltip'
+import { Tutorials } from './components/Tutorials'
 import { TutorialsSlider } from './components/TutorialsSlider'
 import { VisitLibrary } from './components/VisitLibrary'
 import { WorkableOpenRoles } from './components/WorkableOpenRoles'
@@ -201,6 +202,7 @@ export const shortcodes = {
     TableOfContents,
     TeamQuote,
     Tooltip,
+    Tutorials,
     TutorialsSlider,
     VisitLibrary,
     WorkableOpenRoles,
